Fixed openssl 3.0 incompatible error

+++ src/utils/jws_utils/src/jws_utils.c
@@ -555,12 +555,13 @@ JWSResult IsSigningKeyDisallowed(const char* sjwkJsonStr, VECTOR_HANDLE disabled
     }
 
 #ifdef TRACE_DISABLED_SIGNING_KEY
-    char* base64urlSha256HashPubKey = Base64URLEncode(CONSTBUFFER_GetContent(sha256HashPubKey)->buffer, CONSTBUFFER_GetContent(sha256HashPubKey)->size);
+    char* base64urlSha256HashPubKey = Base64URLEncode(
+        CONSTBUFFER_GetContent(sha256HashPubKey)->buffer, CONSTBUFFER_GetContent(sha256HashPubKey)->size);
     printf("base64url encoding of sha256 hash of public key: %s\n", base64urlSha256HashPubKey);
 #endif
 
     // See if the hash of public key is on the Disallowed List.
-    for (int i = 0; i < VECTOR_size(disabledHashOfPubKeysList); ++i)
+    for (size_t i = 0; i < VECTOR_size(disabledHashOfPubKeysList); ++i)
     {
         const ADUC_RootKeyPackage_Hash* DisallowedEntry =
             (ADUC_RootKeyPackage_Hash*)VECTOR_element(disabledHashOfPubKeysList, i);
